feat(js): Com 229 update the in app preview component in the web app to (#6600)
