+++ packages/js/src/ui/components/elements/Markdown.tsx
@@ -1,43 +1,7 @@
 import { createMemo, For, JSX } from 'solid-js';
 import { useStyle } from '../../helpers';
+import { parseMarkdownIntoTokens } from '../../internal';
 import { AppearanceKey } from '../../types';
-
-interface Token {
-  type: 'bold' | 'text';
-  content: string;
-}
-
-const parseMarkdownIntoTokens = (text: string): Token[] => {
-  const tokens: Token[] = [];
-  let buffer = '';
-  let inBold = false;
-
-  for (let i = 0; i < text.length; i += 1) {
-    // Check if it's an escaped character
-    if (text[i] === '\\' && text[i + 1] === '*') {
-      buffer += '*';
-      i += 1;
-    }
-    // Check for bold marker **
-    else if (text[i] === '*' && text[i + 1] === '*') {
-      if (buffer) {
-        tokens.push({ type: inBold ? 'bold' : 'text', content: buffer });
-        buffer = '';
-      }
-      inBold = !inBold;
-      i += 1;
-    } else {
-      buffer += text[i];
-    }
-  }
-
-  // Push any remaining buffered text as a token
-  if (buffer) {
-    tokens.push({ type: inBold ? 'bold' : 'text', content: buffer });
-  }
-
-  return tokens;
-};
 
 const Bold = (props: { children?: JSX.Element; appearanceKey?: AppearanceKey }) => {
   const style = useStyle();
